requirements.txt
# Core Telegram Bot Framework
python-telegram-bot==20.3
asyncio-contextmanager==1.0.1
httpx==0.24.1
aiohttp==3.9.1

# LLM & AI (Groq)
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.1.0
groq==0.4.1

# Data & Validation
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# API Clients  
requests==2.31.0
praw==7.7.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
